Organization and Suggestive Styling
Simply add a organization system to make it more accessible, plus added some styling to make it easier to look at it. DO NOT MERGE YET!!
