--- src/App.css
@@ -1,72 +1,0 @@
-/* App.css */
-/*
-#page-container {
-  padding-top:3%;
-  padding-bottom:5%;
-}
-*/
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  background-color: #f0f0f0;
-}
-
-.App {
-text-align: center;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: 24px; /* Adjust the font size as needed */
-  color: white;
-}
-
-.App-link {
-color: #61dafb;
-}
-
-nav {
-  background-color: #333;
-  padding: 10px;
-  text-align: center;
-}
-
-.App nav{
-  width: 100%;
-  height: 5%;
-}
-
-#Header{
-  position:fixed;
-  width:100%;
-  height:3%;
-  /*background-color:#ffd700;*/
-  background-color:rgba(0,100,100,.5);
-  color:#000080;
-  font-size:100%;
-  top:0;
-}
-
-#Footer {
-  position:fixed;
-  width:100%;
-  height:5%;
-  bottom:0;
-  /*background-color: #000080;*/
-  background-color:rgba(0,100,100,.5);
-}
-
-.navButton{
-  color:blue;
-  background-color: #ffd700;
-}
-
-#SearchBar {
-  display:inline-block;
-}
