+++ src/Main/App.js
@@ -1,15 +1,14 @@
 // App.js
-import React from 'react';
-import './App.css'; // Import the CSS file
+import React from 'react';// Import the CSS file
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
 //import all components
-import Home from './components/home';
-import News from './components/news';
-import Help from './components/help';
-import Settings from './components/settings';
-import Sports from './components/sports';
-import Footer from './components/footer';
-import Header from './components/header';
+import Home from '../components/screens/home';
+import News from '../components/screens/news';
+import Help from '../components/screens/help';
+import Settings from '../components/screens/settings';
+import Sports from '../components/screens/sports';
+import Footer from '../components/Hooks/Constants/footer';
+import Header from '../components/Hooks/Constants/header';
 
 
 const App = () => {
