+++ src/components/Hooks/Constants/FHstyles.css
@@ -1,0 +1,59 @@
+/* styles.css */
+
+/* Global styles */
+body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+}
+
+/* Navigation styles */
+#Footer {
+    background-color: navy;
+    padding: 10px;
+}
+
+.navButton {
+    background-color: navy;
+    color: yellow;
+    border: none;
+    padding: 8px 16px;
+    cursor: pointer;
+    font-size: 16px;
+    margin: 0;
+    width: calc(100% / 5); /* Divide the width equally among the buttons */
+}
+
+.navButton:hover {
+    background-color: yellow;
+    color: navy;
+}
+
+/* SearchBar styles */
+#SearchBar {
+    display: flex;
+    align-items: center;
+}
+
+.search-bar-input {
+    padding: 8px;
+    border: none;
+    border-radius: 4px;
+    margin-right: 8px;
+    font-size: 16px;
+    background-color: white; /* Background color for input */
+}
+
+.search-bar-button {
+    padding: 8px 16px;
+    background-color: yellow;
+    color: navy;
+    border: none;
+    border-radius: 4px;
+    font-size: 16px;
+    cursor: pointer;
+}
+
+.search-bar-button:hover {
+    background-color: #ffeb3b; /* Hover color for button */
+}
