+++ src/components/Hooks/Constants/header.js
@@ -1,18 +1,20 @@
 import React from 'react';
-import SearchBar from "./searchBar";
+import SearchBar from '../Search/searchBar';
+import './FHstyles.css'; // Import the CSS stylesheet
 
-const Footer = () =>{
+const Header = () => {
     // Function to handle search logic
     const handleSearch = (searchTerm) => {
         // Perform search logic here (if needed)
-        console.log("Search term:", searchTerm);
+        console.log('Search term:', searchTerm);
     };
+
     return(
         <nav id="Header">
-            <p style={{display:'inline'}}>MSU News</p>
+            <p className="headerTitle">MSU News</p>
             <SearchBar id="SearchBar" onSearch={handleSearch}/>
         </nav>
     );
 }
 
-export default Footer;
+export default Header;
