+++ src/components/Hooks/Search/searchBar.css
@@ -1,0 +1,28 @@
+/* SearchBar.css */
+.search-bar-form {
+  display: flex;
+  align-items: center;
+}
+
+.search-bar-input {
+  padding: 8px;
+  border: 1px solid navy;
+  border-radius: 4px;
+  margin-right: 8px;
+  font-size: 16px;
+  flex: 1;
+}
+
+.search-bar-button {
+  padding: 8px 16px;
+  background-color: navy;
+  color: yellow;
+  border: none;
+  border-radius: 4px;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+.search-bar-button:hover {
+  background-color: #001f3f; /* Darker shade of navy for hover effect */
+}
