--- src/index.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Your News App</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            padding: 0;
-            background-color: #f0f0f0;
-        }
-
-        header {
-            background-color: #222;
-            color: white;
-            padding: 10px;
-            text-align: center;
-        }
-
-        nav {
-            background-color: #333;
-            padding: 10px;
-            text-align: center;
-        }
-
-        nav a {
-            color: white;
-            text-decoration: none;
-            margin: 0 15px;
-            font-weight: bold;
-        }
-
-        main {
-            padding: 20px;
-        }
-
-        section {
-            margin-bottom: 20px;
-        }
-
-        footer {
-            background-color: #222;
-            color: white;
-            padding: 10px;
-            text-align: center;
-        }
-    </style>
-</head>
-<body>
-    <header>
-        <h1>Your News App</h1>
-    </header>
-    <nav>
-        <a href="#">Home</a>
-        <a href="#">Politics</a>
-        <a href="#">Business</a>
-        <a href="#">Technology</a>
-        <a href="#">Entertainment</a>
-    </nav>
-    <main>
-        <section>
-            <!-- News articles go here -->
-            <h2>Breaking News</h2>
-            <p>Article 1 summary...</p>
-            <p>Article 2 summary...</p>
-            <!-- Add more articles as needed -->
-        </section>
-    </main>
-    <footer>
-        &copy; 2024 Your News App. All rights reserved.
-    </footer>
-</body>
-</html>
